refactor(mcp): extract flight JSON-schema builders into schema_flights.go
tools_flights.go was 808 LOC — over the 800-LOC file-hygiene gate
(workflow-hygiene / check-file-size.sh) and a responsibility collage of
schema literals + tool registration + orchestration. Move the two pure JSON
Schema builders (flightSearchOutputSchema, dateSearchOutputSchema) verbatim
into a new same-package file mcp/schema_flights.go, along the schema/
orchestration seam. Pure move: no renames, no signature or behaviour change;
callers unchanged (same package). tools_flights.go: 808 -> 661 LOC.

Implemented by grok 4.5; reviewed by Claude. go build + go test ./mcp green.
